Update with review suggestions
