+++ frontend/awx/views/schedules/hooks/scheduleHelpers.tsx
@@ -1,5 +1,5 @@
 import { DateTime } from 'luxon';
-import { ReactElement, useCallback } from 'react';
+import { ReactElement } from 'react';
 import { useWatch } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { useLocation, useParams } from 'react-router-dom';
@@ -22,11 +22,10 @@ import { InventorySource } from '../../../interfaces/InventorySource';
 import { JobTemplate } from '../../../interfaces/JobTemplate';
 import { LaunchConfiguration } from '../../../interfaces/LaunchConfiguration';
 import { Project } from '../../../interfaces/Project';
-import { ScheduleFormWizard, schedulePageUrl } from '../types';
+import { ScheduleFormWizard } from '../types';
 import { WorkflowJobTemplate } from '../../../interfaces/WorkflowJobTemplate';
 import { AwxRoute } from '../../../main/AwxRoutes';
 import { PageFormInventorySelect } from '../../../resources/inventories/components/PageFormInventorySelect';
-import { Schedule } from '../../../interfaces/Schedule';
 
 export const resourceEndPoints: { [key: string]: string } = {
   inventories: awxAPI`/inventory_sources/`,
@@ -40,84 +39,6 @@ export const scheduleResourceTypeOptions: string[] = [
   'inventory',
   'projects',
 ];
-
-export function useGetScheduleUrl() {
-  const { t } = useTranslation();
-  return useCallback(
-    (route: string, schedule: Schedule) => {
-      const unified_job_type = schedule.summary_fields.unified_job_template.unified_job_type;
-      const params = {
-        id: schedule.summary_fields.unified_job_template?.id.toString(),
-        schedule_id: schedule.id.toString(),
-      };
-
-      if (unified_job_type === 'inventory_update') {
-        const inventory_update_params = {
-          id: schedule.summary_fields.inventory?.id.toString(),
-          source_id: schedule.summary_fields.unified_job_template.id.toString(),
-          schedule_id: schedule.id.toString(),
-        };
-        return {
-          name: t('Inventory sync'),
-          details: {
-            pageId: AwxRoute.InventorySourceScheduleDetails,
-            params: inventory_update_params,
-          },
-          create: {
-            pageId: AwxRoute.InventorySourceScheduleCreate,
-            params: inventory_update_params,
-          },
-          edit: {
-            pageId: AwxRoute.InventorySourceScheduleEdit,
-            params: inventory_update_params,
-          },
-          resource: {
-            pageId: AwxRoute.InventorySourceDetail,
-            params: inventory_update_params,
-          },
-        }[route] as string | schedulePageUrl;
-      }
-      if (unified_job_type === 'job') {
-        return {
-          name: t('Playbook run'),
-          details: { pageId: AwxRoute.JobTemplateScheduleDetails, params },
-          create: { pageId: AwxRoute.JobTemplateScheduleCreate, params },
-          edit: { pageId: AwxRoute.JobTemplateScheduleEdit, params },
-          resource: { pageId: AwxRoute.JobTemplateDetails, params },
-        }[route] as string | schedulePageUrl;
-      }
-      if (unified_job_type === 'project_update') {
-        return {
-          name: t('Source control update'),
-          details: { pageId: AwxRoute.ProjectScheduleDetails, params },
-          create: { pageId: AwxRoute.ProjectScheduleCreate, params },
-          edit: { pageId: AwxRoute.ProjectScheduleEdit, params },
-          resource: { pageId: AwxRoute.ProjectDetails, params },
-        }[route] as string | schedulePageUrl;
-      }
-      if (unified_job_type === 'system_job') {
-        return {
-          name: t('Management job'),
-          details: { pageId: AwxRoute.ManagementJobScheduleDetails, params },
-          create: { pageId: AwxRoute.ManagementJobScheduleCreate, params },
-          edit: { pageId: AwxRoute.ManagementJobScheduleEdit, params },
-          resource: { pageId: AwxRoute.ManagementJobDetails, params },
-        }[route] as string | schedulePageUrl;
-      }
-      if (unified_job_type === 'workflow_job') {
-        return {
-          name: t('Workflow job'),
-          details: { pageId: AwxRoute.WorkflowJobTemplateScheduleDetails, params },
-          create: { pageId: AwxRoute.WorkflowJobTemplateScheduleCreate, params },
-          edit: { pageId: AwxRoute.WorkflowJobTemplateScheduleEdit, params },
-          resource: { pageId: AwxRoute.WorkflowJobTemplateDetails, params },
-        }[route] as string | schedulePageUrl;
-      }
-      return '';
-    },
-    [t]
-  );
-}
 
 export function useGetSchedulCreateUrl(sublistEndPoint?: string) {
   const getPageUrl = useGetPageUrl();
